수정: Homepage Components 분리 및 Refactoring
